Move build_project to the base validator
