+++ spectacles/validators/content.py
@@ -1,9 +1,7 @@
 from typing import List, Optional, Any, Dict
 from spectacles.validators.validator import Validator
 from spectacles.client import LookerClient
-from spectacles.lookml import Project, Model
-from spectacles.exceptions import LookMlNotFound, ContentError
-from spectacles.select import is_selected
+from spectacles.exceptions import ContentError
 from spectacles.lookml import Explore
 from spectacles.logger import GLOBAL_LOGGER as logger
 
@@ -12,8 +10,7 @@ class ContentValidator(Validator):
     def __init__(
         self, client: LookerClient, project: str, exclude_personal: bool = False
     ):
-        super().__init__(client)
-        self.project = Project(project, models=[])
+        super().__init__(client, project)
         personal_folders = self._get_personal_folders() if exclude_personal else []
         self.personal_folders: List[int] = personal_folders
 
@@ -24,54 +21,6 @@ class ContentValidator(Validator):
             if folder["is_personal"] or folder["is_personal_descendant"]:
                 personal_folders.append(folder["id"])
         return personal_folders
-
-    def build_project(
-        self,
-        selectors: Optional[List[str]] = None,
-        exclusions: Optional[List[str]] = None,
-    ) -> None:
-        """Creates an object representation of the project's LookML.
-
-        Args:
-            selectors: List of selector strings in 'model_name/explore_name' format.
-                The '*' wildcard selects all models or explores. For instance,
-                'model_name/*' would select all explores in the 'model_name' model.
-
-        """
-        # Assign default values for selectors and exclusions
-        if selectors is None:
-            selectors = ["*/*"]
-        if exclusions is None:
-            exclusions = []
-
-        all_models = [
-            Model.from_json(model) for model in self.client.get_lookml_models()
-        ]
-        project_models = [
-            model for model in all_models if model.project_name == self.project.name
-        ]
-
-        if not project_models:
-            raise LookMlNotFound(
-                name="project-models-not-found",
-                title="No configured models found for the specified project.",
-                detail=(
-                    f"Go to {self.client.base_url}/projects and confirm "
-                    "a) at least one model exists for the project and "
-                    "b) it has an active configuration."
-                ),
-            )
-
-        for model in project_models:
-            model.explores = [
-                explore
-                for explore in model.explores
-                if is_selected(model.name, explore.name, selectors, exclusions)
-            ]
-
-        self.project.models = [
-            model for model in project_models if len(model.explores) > 0
-        ]
 
     def validate(self) -> Dict[str, Any]:
         result = self.client.content_validation()
